added compatability to valve control
